Moved to the test list to its own class.
